# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crafted By API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/business": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Get a list of business",
                "description": "Display a listing of the resource.",
                "operationId": "45afc1ca4cf6872e989e121f5305d8d8",
                "responses": {
                    "200": {
                        "description": "Successful opération",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BusinessCard"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Business"
                ],
                "summary": "Business Store",
                "description": "Create Business",
                "operationId": "addBusiness",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BusinessStore"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessCard"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Invalid Input"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/business/{business_id}": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Find a business by Id",
                "description": "Return a single business",
                "operationId": "getBusinessById",
                "parameters": [
                    {
                        "name": "business_id",
                        "in": "path",
                        "description": "Id of business to return",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Business"
                ],
                "summary": "Update existing business",
                "description": "Returns updated business data",
                "operationId": "updateBusiness",
                "parameters": [
                    {
                        "name": "business_id",
                        "in": "path",
                        "description": "Business id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BusinessModel"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Business"
                ],
                "summary": "Delete existing business",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteBusiness",
                "parameters": [
                    {
                        "name": "business_id",
                        "in": "path",
                        "description": "Business id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BusinessCard": {
                "title": "Business Card",
                "description": "Small Business Data",
                "properties": {
                    "Id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "uuid"
                    },
                    "name": {
                        "title": "Business name",
                        "type": "string"
                    },
                    "email": {
                        "title": "Business email",
                        "type": "string"
                    },
                    "theme": {
                        "$ref": "#/components/schemas/Theme"
                    }
                },
                "type": "object"
            },
            "BusinessStore": {
                "title": "Business Store",
                "description": "Business model to Store",
                "properties": {
                    "Id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "uuid"
                    },
                    "name": {
                        "title": "Business name",
                        "type": "string"
                    },
                    "email": {
                        "title": "Business email",
                        "type": "string",
                        "example": "business@business.com"
                    },
                    "theme": {
                        "$ref": "#/components/schemas/Theme"
                    },
                    "description": {
                        "title": "Business description",
                        "type": "string"
                    },
                    "history": {
                        "title": "Business history",
                        "type": "string"
                    },
                    "address": {
                        "title": "Business address",
                        "type": "string",
                        "example": "Papeteries"
                    },
                    "zip_code": {
                        "title": "Business post code",
                        "type": "string",
                        "example": "74960"
                    },
                    "city": {
                        "title": "Business city location",
                        "type": "string",
                        "example": "Cran-Gevrier"
                    },
                    "logo": {
                        "title": "Business logo",
                        "type": "string"
                    },
                    "specialities": {
                        "title": "Business specialities",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "BusinessModel": {
                "title": "Business Model",
                "properties": {
                    "Id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "uuid"
                    },
                    "name": {
                        "title": "Business name",
                        "type": "string"
                    },
                    "email": {
                        "title": "Business email",
                        "type": "string"
                    },
                    "theme": {
                        "$ref": "#/components/schemas/Theme"
                    },
                    "owners": {
                        "title": "Business Owners",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserCard"
                        }
                    },
                    "description": {
                        "title": "Business description",
                        "type": "string"
                    },
                    "history": {
                        "title": "Business history",
                        "type": "string"
                    },
                    "address": {
                        "title": "Business address",
                        "type": "string"
                    },
                    "zip_code": {
                        "title": "Business post code",
                        "type": "string"
                    },
                    "city": {
                        "title": "Business city location",
                        "type": "string"
                    },
                    "logo": {
                        "title": "Business logo",
                        "type": "string"
                    },
                    "specialities": {
                        "title": "Business specialities",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "products": {
                        "title": "Business products",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductCard"
                        }
                    },
                    "created_at": {
                        "title": "Business registering date",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "ProductCard": {
                "title": "ProductCard",
                "description": "Product Model for cards",
                "properties": {
                    "Id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "uuid"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Product name",
                        "type": "string",
                        "example": "Chaise en bois rare"
                    },
                    "image": {
                        "title": "Image",
                        "description": "Image URL",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "description": "Product price",
                        "type": "number",
                        "format": "float",
                        "example": "123.45"
                    },
                    "active": {
                        "title": "Active",
                        "description": "Is product available to sell",
                        "type": "boolean"
                    },
                    "business": {
                        "$ref": "#/components/schemas/BusinessCard"
                    }
                },
                "type": "object"
            },
...
            "Theme": {
                "title": "Business Theme",
                "properties": {
                    "layer": {
                        "title": "Theme Layer",
                        "type": "string"
                    },
                    "color1": {
                        "title": "Theme color n°1",
                        "type": "string"
                    },
                    "color2": {
                        "title": "Theme color n°2",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserCard": {
                "title": "User card",
                "description": "Small User data",
                "properties": {
                    "Id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "uuid"
                    },
                    "lastname": {
                        "title": "User Lastname",
                        "type": "string"
                    },
                    "firstname": {
                        "title": "User firstname",
                        "type": "string"
                    },
                    "role": {
                        "title": "User role",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}